feat: show a welcome view if playgrounds list is empty
